Move lfp_reader to io/, validate electrodes early in configuration
- Move neurodamus/lfp_reader.py to neurodamus/io/lfp_reader.py
- Add @SimConfig.validator to check LFP electrodes files at config time
- Add disable_reports to CliOptions so validator can skip when reports disabled
- Replace ConfigurationError with LFPFileValidationError (breaks circular import)
- Update docs to show electrodes_file in report block (not run section)
